requirements.txt
python-telegram-bot==20.7
twilio==8.12.0
python-dotenv==1.0.0
SQLAlchemy==2.0.27
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
aiohttp==3.9.3 